Add .env file handling to dotnet-build workflow
This commit introduces a step to copy the .env file to the publish output directory in the dotnet-build.yml workflow. The .env file is also included in the list of artifacts to be uploaded, ensuring it is available alongside other published files like docker-compose.yml and aspnetapp.pfx.

Update build and configuration for logging and networking

- Added step to copy `.env` file in `dotnet-build.yml` and included it in uploaded artifacts.
- Introduced `Serilog.Enrichers.Environment` and `Serilog.Exceptions` packages in `PropertyManagement.Web.csproj` for enhanced logging.
- Updated `docker-compose.yml` to replace `appnet` with a new network `mynet`, affecting service connectivity.
- Adjusted `elasticsearch` and `kibana` services to utilize the new `mynet` network.